* commit 'c96d64e7e142d43db1864da17520d69268627aa3':
  invalid type cast correction
  list_users tipo correction

+++ src/mod/endpoints/mod_sofia/sofia_reg.c
@@ -857,7 +857,7 @@ long sofia_reg_uniform_distribution(int max)
 	int result;
 	int range = max + 1;
 
-	srand((unsigned) switch_thread_self() + switch_micro_time_now());
+	srand((intptr_t) switch_thread_self() + switch_micro_time_now());
 	result = (int)((double)rand() / (((double)RAND_MAX + (double)1) / range));
 
 	switch_log_printf(SWITCH_CHANNEL_LOG, SWITCH_LOG_DEBUG9, "Generated random %ld, max is %d\n", (long) result, max);
